transfered getReachTimeToRegion to TsunamiRegionClass and revised TsunamiParserClass
