+++ src/main/java/jp/ac/aiit/pbl/disaster/tsunami/TsunamiParser.java
@@ -25,50 +25,43 @@ public class TsunamiParser implements DisasterParser{
         
         List<TsunamiRegion> tsunamiRegions = new ArrayList<>();
         
-        ExpectedTsunamiArrivalTime tsunamiForecast = new ExpectedTsunamiArrivalTime();
-        
         tsunamiRegions.add(new TsunamiRegion(
-                tsunamiForecast.getReachTimeToRegion(
-                        tsunami.getPrefix().getReportTime(), //IssueDate
-                        Integer.parseInt(qzMessage.substring(84, 85), 2), //day
-                        Integer.parseInt(qzMessage.substring(85, 90),2),  //hour
-                        Integer.parseInt(qzMessage.substring(90, 96), 2)), //minute
+                tsunami.getPrefix().getReportTime(), //IssueDate
+                Integer.parseInt(qzMessage.substring(84, 85), 2), //day
+                Integer.parseInt(qzMessage.substring(85, 90),2),  //hour
+                Integer.parseInt(qzMessage.substring(90, 96), 2), //minute
                 Integer.parseInt(qzMessage.substring(96, 100), 2),
                 Integer.parseInt(qzMessage.substring(100, 110),2)));
         
         tsunamiRegions.add(new TsunamiRegion(
-                tsunamiForecast.getReachTimeToRegion(
-                        tsunami.getPrefix().getReportTime(),
-                        Integer.parseInt(qzMessage.substring(110, 111), 2), //day
-                        Integer.parseInt(qzMessage.substring(111, 116),2),  //hour
-                        Integer.parseInt(qzMessage.substring(116, 122), 2)), //minute
+                tsunami.getPrefix().getReportTime(), //IssueDate
+                Integer.parseInt(qzMessage.substring(110, 111), 2), //day
+                Integer.parseInt(qzMessage.substring(111, 116),2),  //hour
+                Integer.parseInt(qzMessage.substring(116, 122), 2), //minute
                 Integer.parseInt(qzMessage.substring(122, 126), 2),
                 Integer.parseInt(qzMessage.substring(126, 136),2)));
         
         tsunamiRegions.add(new TsunamiRegion(
-                tsunamiForecast.getReachTimeToRegion(
-                        tsunami.getPrefix().getReportTime(),
-                        Integer.parseInt(qzMessage.substring(136, 137), 2), //day
-                        Integer.parseInt(qzMessage.substring(137, 142),2),  //hour
-                        Integer.parseInt(qzMessage.substring(142, 148), 2)), //minute
+                tsunami.getPrefix().getReportTime(), //IssueDate
+                Integer.parseInt(qzMessage.substring(136, 137), 2), //day
+                Integer.parseInt(qzMessage.substring(137, 142),2),  //hour
+                Integer.parseInt(qzMessage.substring(142, 148), 2), //minute
                 Integer.parseInt(qzMessage.substring(148, 152), 2),
                 Integer.parseInt(qzMessage.substring(152, 162),2)));
         
         tsunamiRegions.add(new TsunamiRegion(
-                tsunamiForecast.getReachTimeToRegion(
-                        tsunami.getPrefix().getReportTime(),
-                        Integer.parseInt(qzMessage.substring(162, 163), 2), //day
-                        Integer.parseInt(qzMessage.substring(163, 168),2),  //hour
-                        Integer.parseInt(qzMessage.substring(168, 174), 2)), //minute
+                tsunami.getPrefix().getReportTime(), //IssueDate
+                Integer.parseInt(qzMessage.substring(162, 163), 2), //day
+                Integer.parseInt(qzMessage.substring(163, 168),2),  //hour
+                Integer.parseInt(qzMessage.substring(168, 174), 2), //minute
                 Integer.parseInt(qzMessage.substring(174, 178), 2),
                 Integer.parseInt(qzMessage.substring(178, 188),2)));
         
         tsunamiRegions.add(new TsunamiRegion(
-                tsunamiForecast.getReachTimeToRegion(
-                        tsunami.getPrefix().getReportTime(),
-                        Integer.parseInt(qzMessage.substring(188, 189), 2), //day
-                        Integer.parseInt(qzMessage.substring(189, 194),2),  //hour
-                        Integer.parseInt(qzMessage.substring(194, 200), 2)), //minute
+                tsunami.getPrefix().getReportTime(), //IssueDate
+                Integer.parseInt(qzMessage.substring(188, 189), 2), //day
+                Integer.parseInt(qzMessage.substring(189, 194),2),  //hour
+                Integer.parseInt(qzMessage.substring(194, 200), 2), //minute
                 Integer.parseInt(qzMessage.substring(200, 204), 2),
                 Integer.parseInt(qzMessage.substring(204, 214),2)));
         tsunami.setTsunamiRegions(tsunamiRegions);
